backend/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
asgiref==3.3.1
    # via django
django-cors-headers==3.7.0
    # via -r requirements.in
django-extensions==3.1.1
    # via -r requirements.in
django==3.1.6
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   djangorestframework
djangorestframework==3.12.2
    # via -r requirements.in
pillow==8.1.0
    # via -r requirements.in
python-dateutil==2.8.1
    # via -r requirements.in
python-dotenv==0.15.0
    # via -r requirements.in
pytz==2021.1
    # via django
six==1.15.0
    # via python-dateutil
sqlparse==0.4.1
    # via django
